requirements.txt
beautifulsoup4==4.9.1
firebase-admin==4.3.0
google-api-python-client==1.9.3
requests==2.23.0
python-telegram-bot==12.7
python-dotenv==0.13.0
pymongo==3.10.1
dnspython==1.16.0